+++ Sources/Request Strategies/User Clients/FetchUserClientsActionHandler.swift
@@ -124,14 +124,3 @@ final class FetchUserClientsActionHandler: ActionHandler<FetchUserClientsAction>
     }
 
 }
-
-extension ZMTransportResponse {
-
-    var errorInfo: (status: Int, label: String, message: String) {
-        let payload = self.payload?.asDictionary()
-        let label = payload?["label"] as? String
-        let message = payload?["message"] as? String
-        return (httpStatus, label ?? "?", message ?? "?")
-    }
-
-}
